perf(model): 设置模型使用 float32 类型
- 在加载模型时指定 torch_dtype=torch.float32，以确保模型使用浮点数精度
- 这种修改可以提高模型的运行速度和减少内存消耗

+++ utils/images_vector/model_factory.py
@@ -63,7 +63,8 @@ class ImageVectorizer(ABC):
             with cls.lock:
                 if cls.model is None:  # 内部检查，防止多个线程竞争
                     model_path = cls.get_model_path()
-                    cls.model = AutoModelForZeroShotImageClassification.from_pretrained(model_path)
+                    cls.model = AutoModelForZeroShotImageClassification.from_pretrained(model_path,
+                                                                                        torch_dtype=torch.float32)
                     cls.model.to(cls.get_device())
         return cls.model
